Add a way to "sync" issue body/assignee only, by a query
