+++ src/main/java/org/hibernate/infra/replicate/jira/service/jira/handler/JiraIssueSimpleUpsertEventHandler.java
@@ -4,21 +4,21 @@ import org.hibernate.infra.replicate.jira.service.jira.HandlerProjectContext;
 import org.hibernate.infra.replicate.jira.service.jira.model.rest.JiraIssue;
 import org.hibernate.infra.replicate.jira.service.reporting.ReportingConfig;
 
-public class JiraIssueTransitionOnlyEventHandler extends JiraIssueInternalAbstractEventHandler {
+public class JiraIssueSimpleUpsertEventHandler extends JiraIssueInternalAbstractEventHandler {
 
-	public JiraIssueTransitionOnlyEventHandler(ReportingConfig reportingConfig, HandlerProjectContext context,
+	public JiraIssueSimpleUpsertEventHandler(ReportingConfig reportingConfig, HandlerProjectContext context,
 			JiraIssue issue) {
 		super(reportingConfig, context, issue);
 	}
 
 	@Override
 	protected void updateAction(String destinationKey, JiraIssue sourceIssue) {
-		applyTransition(sourceIssue, destinationKey);
+		updateIssueBody(sourceIssue, destinationKey);
 	}
 
 	@Override
 	public String toString() {
-		return "JiraIssueTransitionOnlyEventHandler[" + "objectId=" + objectId + ", project=" + context.projectName()
+		return "JiraIssueSimpleUpsertEventHandler[" + "objectId=" + objectId + ", project=" + context.projectName()
 				+ ']';
 	}
 }
